### 2021-01-18 #2963 @jedbrown
I'm conflicted. I'd mention that Elemental (written by JOSS editor @poulson) has support for pseudospectra.
