### 2021-01-20 #2963 @kthyng
Hi @jsoishi and thanks for your submission. The overwhelming comment from the editorial board discussing whether this was in scope or not was "why isn't this part of Dedalus?"
