### 2021-01-20 #2963 @jsoishi
Hi @kthyng,

Thanks for considering the paper. We feel that this question is best answered by considering the scope of the two packages. Dedalus is dedicated to the spectral discretization and solution of PDEs. It provides basic solvers but intentionally does not implement high-level workflows that are not inherently tied to the spectral discretization. For example, it does not do any degree of spatial refinement or error checking of solutions. It does not include higher level functionality like adjoint looping for PDE optimization, rather just the basic solvers for implementing all of these tasks.

Eigentools focuses on implementing well-tested algorithms for error estimation, automated parameter space exploration and optimization, tools for plotting, and other high-level functionality specific to eigenvalue problems. While it currently interfaces with Dedalus, nothing in eigentools inherently requires spectral discretization. The tools for eigenvalue problems provided by eigentools are therefore not within the scope of the core Dedalus code.

As a similar example of separating scope, the [dolfin-adjoint solver (published in JOSS)](https://joss.theoj.org/papers/10.21105/joss.01292) constructs adjoint solvers for previously derived FENICS PDE models. Adjoint methods are also related to the pseudospectra implemented by eigentools, as they provide similar information about the underlying PDE models.

Separately, having read the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort), we feel that we meet all of the criteria:

- eigentools is five years old with substantial development effort over that time (first commit 15 Jan 2016)
- eigentools has 146 commits by 5 authors
- eigentools has 1148 SLOC and 1393 comment lines; a core development philosophy is code compactness
- eigentools has been cited by at least 7 papers, including at least one by an author unconnected to the development team
- eigentools is in active use by several research groups world wide; once a JOSS paper appears, these groups are highly likely to cite it
- eigentools is feature-complete; it implements algorithms distinct from those provided by Dedalus
- eigentools is not a single-function package, nor an API client.

We hope this information makes clear why we developed eigentools in this way and why we believe it is appropriate for publication in JOSS.

Sincerely,

@jsoishi
on behalf of the authors
