### 2021-01-22 #2963 @kthyng
Hi @jsoishi, thanks for your response. The editorial board recommends moving forward to review on this submission. One additional note to leave you with was that the software could be more useful if it supported models defined outside of Dedalus (similar to how scipy.integrate is general-purpose).
