### 2021-01-22 DedalusProject/eigentools#5 @dpsanders
ReadTheDocs link in README is broken
Part of JOSS pre-review at https://github.com/openjournals/joss-reviews/issues/2963
