### 2021-01-22 3bbded2
Fixed Read the Docs link
Closes issue #5
