### 2021-01-22 DedalusProject/eigentools#5 @jsoishi
Fixed with commit 3bbded2.
