### 2021-01-22 #2963 @dpsanders
👋 @jsoishi: Thanks for submitting your work to JOSS!

My first initial observations:
- I don't see any narrative documentation at all that explains how to use the package.
- As far as I can tell from the README, `eigentools` requires `dedalus` to be installed. In that case, `pip install eigentools` should automatically install `dedalus`.
- Could you please address @kthyng's comment that the package should ideally work beyond the scope of `dedalus`.

Thanks!
