### 2021-02-01 780e8c8
Documentation update
* added intro text
* links to example notebooks for each feature
* updated installation instructions
