### 2021-02-01 b841c9e
Restructuring of example notebooks
* specific features are broken out by sections
