+++ docs/index.rst
@@ -3,14 +3,15 @@ Eigentools
 
 Eigentools is a set of tools for studying linear eigenvalue problems. The underlying eigenproblems are solved using `Dedalus <http://dedalus-project.org>`_, which provides a domain-specific language for partial differential equations. Each entry in the following list of features links to a Jupyter notebook giving an example of its use.
 
-* :ref:`automatic rejection of unresolved eigenvalues </notebooks/Orr Somerfeld pseudospectra.ipynb>`
-* :ref:`simple plotting of specified eigenmodes </notebooks/Mixed Layer Instability.ipynb>`
-* :ref:`simple plotting of spectra </notebooks/Orr Somerfeld pseudospectra.ipynb>`
-* :ref:`computation of pseudospectra for any Differential-Algebraic Equations with user-specifiable norms </notebooks/Orr Somerfeld pseudospectra.ipynb>`
-* :ref:`tools to find critical parameters for linear stability analysis </notebooks/Mixed Layer Instability.ipynb>`
-* :ref:`ability to project eigenmode onto 2- or 3-D domain for visualization </notebooks/Mixed Layer Instability.ipynb>`
-* :ref:`ability to output projected eigenmodes as Dedalus-formatted HDF5 file to be used as initial conditions for Initial Value Problems </notebooks/Mixed Layer Instability.ipynb>`
-* :ref:`simple plotting of drift ratios (both ordinal and nearest) to evaluate tolerance for eigenvalue rejection </pages/getting_started.rst>`
+* :ref:`automatic rejection of unresolved eigenvalues </notebooks/Orr Somerfeld pseudospectra.ipynb#Eigenmode-rejection>`
+* :ref:`simple plotting of drift ratios (both ordinal and nearest) to evaluate tolerance for eigenvalue rejection </pages/getting_started.rst#Mode-rejection>`
+
+* :ref:`simple plotting of specified eigenmodes </notebooks/Mixed Layer Instability.ipynb#Plotting-eigenmodes>`
+* :ref:`simple plotting of spectra </notebooks/Orr Somerfeld pseudospectra.ipynb#Plotting-Spectra>`
+* :ref:`computation of pseudospectra for any Differential-Algebraic Equations </notebooks/Orr Somerfeld pseudospectra.ipynb#Pseudospectra>` with :ref:`user-specifiable norms </notebooks/Orr Somerfeld pseudospectra.ipynb#Choosing-an-inner-product-and-norm>`
+* :ref:`tools to find critical parameters for linear stability analysis </notebooks/Mixed Layer Instability.ipynb#Finding-critical-parameters>` with :ref:`user-specifiable definitions of growth and stability </notebooks/Mixed Layer Instability.ipynb#Specifying-a-definition-of-stability>`
+* :ref:`ability to project eigenmode onto 2- or 3-D domain for visualization </notebooks/Mixed Layer Instability.ipynb#Projection-onto-higher-dimensional-domains>`
+* :ref:`ability to output projected eigenmodes as Dedalus-formatted HDF5 file to be used as initial conditions for Initial Value Problems </notebooks/Mixed Layer Instability.ipynb#Writing-Dedalus-HDF5-files>`
 
 Contents
 ========
