### 2021-02-02 07a7685
Fixed bug when reject=False
