### 2021-02-02 #2963 @jsoishi
Thanks @dpsanders for your comments.

First, we have updated the documentation to provide specific links to Jupyter notebooks outlining all the features of eigentools. You can see those efforts in either the main branch of the repository or on [readthedocs](https://eigentools.readthedocs.io/en/latest/).

Second, Dedalus is already in `requirements.txt` for eigentools, so pip will attempt to install it. However, because Dedalus itself requires some dependencies that are not pip installable, we have added a warning that without those dependencies the Dedalus pip install will fail. There are detailed instructions and an automated script to install them all via conda on the Dedalus website.

Third, the expansion of eigentools to support matrices produced by other PDE solvers is an interesting issue. While there are no mathematical issues to overcome, there are significant API changes that would need to be implemented in order to create a wrapper level to translate between the quantities needed by eigentools (e.g. eigenvalues and eigenvectors in certain order, grid point values) and the methods required of the external eigenvalue solvers (e.g. parameter inputs, copy constructors for creating higher resolution problems for eigenvalue rejection). This work is an interesting future direction, but currently, eigentools users haven't yet requested support for other solvers. When that time comes, we would welcome contributions and would be happy to work with developers from other PDE solving packages interested in adapting eigentools for their solvers. We would be happy to address this as future work in the JOSS manuscript if the editors or referees would find that useful.

thanks,
@jsoishi
on behalf of the authors
