### 2021-03-02 #3079 @whedon
[REVIEW]: eigentools: A Python package for studying differential eigenvalue problems with an emphasis on robustness
**Submitting author:** @jsoishi (<a href="http://orcid.org/0000-0001-8531-6570">Jeffrey S. Oishi</a>)
**Repository:** <a href="https://github.com/DedalusProject/eigentools" target ="_blank">https://github.com/DedalusProject/eigentools</a>
**Version:** 2.2106
