### 2021-03-29 #3079 @caropen
I was able to run the software and all provided examples. I could check off all items except the one asking for "Community guidelines", which I could not find. Am I supposed to raise an issue in the original repository? (first time reviewer)
