### 2021-03-30 bf22e4c
Added CONTRIBUTING.md addressing issue #7.
