+++ CONTRIBUTING.md
@@ -1,0 +1,18 @@
+# Contributing to Eigentools #
+
+We welcome contributions, including issue reports, bug fixes, and feature implementations.
+If you would 
+
+## Reporting issues ##
+
+If you find a bug or unexpected behavior, please file an issue report on [github](https://github.com/DedalusProject/eigentools/issues).
+Please provide as much detail as possible, including version of both eigentools and dedalus, platform (Mac/Linux), and a stand alone `.py` file that demonstrates the problem in as simple a manner as possible.
+
+## Proposing features ##
+
+You can propose new features on the issue tracker using the "enhancement" tag.
+
+## Contributing code ##
+
+Code contributions ranging from fixing typos to implementing additional features are most welcome! We use [pull requests](https://github.com/DedalusProject/eigentools/pulls) to integrate contributions into the main codebase. If you would like to contribute and are looking for a place to start, please don't hesitate to contact the authors (emails are in [readme.md](readme.md)).
+
