### 2021-04-05 #3079 @ketch
This looks like a really great package that I myself may use at some point!  Below are the issues I found.

I get some errors when running one of the documentation examples and I raised an issue for that.

Some minor things that should be corrected in the paper:
- line 25: “backgroud”
- line 35: Definition of growth rate is circular
- line 77: “at with”
- Figure 2: 10^-3 factor on color bar is critical but too small.
- Figure 3: “insures” -> “ensures”.
- line 98: “helping to understanding”
- Add a reference for the Orr-Sommerfeld problem.
- Second sentence of contributing.md is incomplete.

Authorship: the package seems to be written almost exclusively by lead author.  What are the contributions of others? (I'm just basing this on Github's commit data)

Installation: the installation instructions at https://eigentools.readthedocs.io/en/latest/pages/installation.html read something like "just do this easy step 1.  All done.  But note that your installation won't actually work because you didn't do step zero.  Now we will tell you about step zero."  I think they should be rewritten more in the form:

Step zero: install Dedalus
Step one: pip install eigentools
Done.
