### 2021-04-19 #3079 @bpbrown
@ketch and JOSS editors,
       I wanted to help answer the authorship question.

We often work in team-coding environments, with several of us simultaneously working on the code while one person makes the logged commits.  Keaton Burns and myself (Ben Brown) jointly worked with Jeff on the refactor and on the JOSS example problems and paper figures, with Jeff Oishi writing and committing the code.  Daniel Lecoanet and Geoffrey Vasil helped with the pseudospectra functionality.  Evan Anders did early work on parameter space searching, laying the foundation for the refactored version.

There are many early commits by both Evan Anders and Susan Clark in the commit history, but these commits are not yet linked to their github identities.  This work had started in a mercurial repo on bitbucket, before transitioning to git on github, and while that history is there, that transition happened after Evan and Susan's primary involvement in the project.  Using "git shortlog --summary --numbered --email", Evan is the second highest contributor by commits after Jeff Oishi, and Susan's commits are also present.  Their github accounts are not yet linked to these older academic e-mails, and we are working with them to resolve that.
