+++ JOSS-paper/paper.bib
@@ -1,3 +1,39 @@
+@article{1993Sci261578T,
+       author = {Trefethen, L. N. and Trefethen, Anne E. and Reddy, Satish C. and Driscoll, Tobin A.},
+        title = "{Hydrodynamic stability without eigenvalues}",
+      journal = {Science},
+     keywords = {Computational Fluid Dynamics, Eigenvalues, Flow Stability, Hydrodynamic Equations, Flow Equations, Flow Geometry, Laminar Flow, Spectral Methods, Transition Flow, Velocity Distribution, Vortices, Fluid Mechanics and Heat Transfer},
+         year = 1993,
+        month = jul,
+       volume = {261},
+       number = {5121},
+        pages = {578-584},
+          doi = {10.1126/science.261.5121.578},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1993Sci...261..578T},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+@article{doi:10.1137/0153002,
+author = {Reddy, Satish C. and Schmid, Peter J. and Henningson, Dan S.},
+title = {Pseudospectra of the Orr–Sommerfeld Operator},
+journal = {SIAM Journal on Applied Mathematics},
+volume = {53},
+number = {1},
+pages = {15-47},
+year = {1993},
+doi = {10.1137/0153002},
+
+URL = { 
+        https://doi.org/10.1137/0153002
+    
+},
+eprint = { 
+        https://doi.org/10.1137/0153002
+    
+}
+}
+
 @article{PhysRevResearch.2.023068,
   title = {Dedalus: A flexible framework for numerical simulations with spectral methods},
   author = {Burns, Keaton J. and Vasil, Geoffrey M. and Oishi, Jeffrey S. and Lecoanet, Daniel and Brown, Benjamin P.},
