+++ JOSS-paper/paper.md
@@ -54,12 +54,12 @@ These analyses require solving eigenvalue problems.
 However, studying these eigenvalues is not without significant peril: discretized PDEs are poorly conditioned and many of the numerical eigenvalues reported by standard solvers are unreliable. Additionally, it is far from trivial to start with a set of PDEs and *construct* a matrix discretization in the first place. In order to solve these problems, we present `eigentools`, a Python package that extends the eigenvalue problem (EVP) capabilities of the Dedalus Project [@PhysRevResearch.2.023068] to provide a complete analysis toolkit for EVPs.
 
 `eigentools` provides a convenient, parallelized interface for both modal and non-modal stability analyses for nearly arbitrary systems of PDEs.
-It provides a toolkit that makes it very easy to take a model, find robust eigenvalues and eigenmodes, and find critical parameter values for stability. The only thing a user needs to do is find a state to linearize about. One constructs a Dedalus `EigenvalueProblem` object for the PDE linearized about the chosen backgroud, and passes that to `eigentools`. `eigentools` provides robust spurious eigenvalue rejection [@boyd2001chebyshev], spectrum and eigenmode visualization, $\varepsilon-$pseudospectra, and the ability to project a given eigenmode onto an 2- or 3-dimensional domain and save it as a Dedalus-formatted HDF5 file to use as an initial condition for a non-linear simulation of the same system. 
+It provides a toolkit that makes it very easy to take a model, find robust eigenvalues and eigenmodes, and find critical parameter values for stability. The only thing a user needs to do is find a state to linearize about. One constructs a Dedalus `EigenvalueProblem` object for the PDE linearized about the chosen background, and passes that to `eigentools`. `eigentools` provides robust spurious eigenvalue rejection [@boyd2001chebyshev], spectrum and eigenmode visualization, $\varepsilon-$pseudospectra, and the ability to project a given eigenmode onto an 2- or 3-dimensional domain and save it as a Dedalus-formatted HDF5 file to use as an initial condition for a non-linear simulation of the same system. 
 
 # Robust Eigenvalues and Finding Critical Parameters
 The core `eigentools` class, `Eigenproblem`, provides a simple interface to accurately solve a Dedalus eigenvalue problem using sparse or dense methods.
-`Eigenproblem` allows the user to choose how "growth" is defined via custom functions of the eigenvalues returned by the solver; this allows the use of positive imaginary parts (e.g. $e^{\sigma t}$, $\sigma \in \mathbb{C}$), negative real parts (e.g. $e^{i(k x - \omega t)}$, $\omega \in \mathbb{C}$), or any other choice.
-The growth rate is defined as the robust eigenvalue with the highest growth rate.
+A common use of eigenvalue problems is stability analysis, in which one characterizes the exponential growth rate $\gamma$ of a mode via (typically) the temporal part of the eigenvalue, $e^{\gamma t}$ with $\gamma \in \mathbb{R}$. When $\gamma > 0$, the mode is unstable. `Eigenproblem` allows the user to choose how this "growth" is defined via custom functions of the eigenvalues returned by the solver. This permits instability to correspond to positive real parts of the eigenvalue (e.g. $e^{\sigma t}$, $\sigma = \gamma + i \omega$ and $\gamma, \omega \in \mathbb{R}$), negative imaginary parts of the eigenvalue (e.g. $e^{i(k x - \omega t)}$, $\omega = \omega_r + i \omega_i$, $\omega_r, \omega_i \in \mathbb{R}$), or any other choice.
[…]
