### 2021-04-27 178909b
Fixed installation instructions per referee suggestion.
