+++ docs/pages/installation.rst
@@ -1,7 +1,9 @@
 Installing eigentools
 *********************
 
-eigentools itself is `pip` installable::
+eigentools requires Dedalus, which you can install via any of the methods found in `the Dedalus installation instructions <https://dedalus-project.readthedocs.io/en/latest/pages/installation.html>`_.
+
+Once Dedalus is installed, eigentools is `pip` installable::
 
   pip install eigentools
 
@@ -10,8 +12,5 @@ If you would like the development version, you can clone the repository and inst
   git clone https://github.com/DedalusProject/eigentools.git
   pip install -e eigentools
 
-**Caution**: eigentools requires Dedalus, which can (and will) also be pip installed, but Dedalus relies on several non-python dependencies that *cannot* be pip installed.
-These must be installed before pip installing eigentools or the process will fail.
-You can install those dependencies following the procedures in `the Dedalus documentation <https://dedalus-project.readthedocs.io/en/latest/>`_.
 
 
