### 2021-04-27 #3079 @jsoishi
@dpsanders @ketch @caropen I've responded to all of @ketch's suggestions. The additions to the paper can be found in the `joss_paper` branch on github; the corrections to the documentation and `contributing.md` are on `master`.

Thanks for the really helpful suggestions.
