### 2021-04-29 #3079 @dpsanders
@jsoishi: Following on from @ketch's comment above, I still find the installation instructions in the repo unclear.

I suggest providing a "quick install" guide in the `eigentools` repository that provides the simplest instructions for installing all required dependencies, and refer to the Dedalus repository for more advanced instructions if necessary.

I am not up to date with the Python packaging situation, but it seems like it should be possible to make

`conda add eigentools`

install *all* required dependencies?
