### 2021-04-30 #3079 @jsoishi
@dpsanders Unfortunately, `conda` has limited support for dependencies requiring locally-built and optimized libraries (e.g. MPI, FFTW provided by supercomputing centers), as Dedalus does. For now, this prevents Dedalus from distributing binaries via `conda` and thus the same for `eigentools`. `eigentools` remains easily `pip` installable as long as Dedalus is properly installed.
