### 2021-05-10 #3079 @dpsanders
@jsoishi: Apologies for the delay.

Here is my attempt to install dedalus following then conda installation instructions linked from the repo.
I first installed miniconda using homebrew, and then installed dedalus with the unmodified conda script.

I get the following errors when I run the tests:

https://gist.github.com/e1a8bb1e691014ee95a0e0a5e96f2c5d
