### 2021-05-10 #3079 @kburns
Hi @dpsanders, this issue has been fixed on the dedalus development branch and is slated to be included in the next release (https://github.com/DedalusProject/dedalus/issues/133).
