### 2021-05-10 #3079 @dpsanders
OK thanks.

Where is the narrative documentation on how to use the package? I thought I saw some but now I can't find it. This should be linked from the GitHub repo.
