### 2021-05-10 #3079 @dpsanders
In particular, not from https://eigentools.readthedocs.io/en/latest/ !
