### 2021-06-03 955f2ad
Fixed H matrix construction; issue warning when ps maxiter reached
* if the iterative solver reached the maximum number of iterations, it
  would throw an IndexError. This commit fixes that behavior.
* if the pseudospectrum calculation reaches maxiter, it will issue a
  warning that the solver did not converge.
