+++ JOSS-paper/paper.bib
@@ -145,7 +145,7 @@ volume = "356",
 pages = "383 - 416",
 year = "2015",
 issn = "0003-4916",
-doi = "https://doi.org/10.1016/j.aop.2015.02.031",
+doi = "10.1016/j.aop.2015.02.031",
 url = "http://www.sciencedirect.com/science/article/pii/S0003491615000901",
 author = "Terry A. Loring",
 keywords = "Topological insulators, -theory, Phase transitions, Algorithms",
@@ -232,18 +232,21 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{2020arXiv201112300L,
+@ARTICLE{2021ApJ.907.64L,
        author = {{Lin}, Min-Kai},
-        title = "{Stratified and vertically-shearing streaming instabilities in protoplanetary disks}",
-      journal = {Astrophysical Journal, accepted},
-     keywords = {Astrophysics - Earth and Planetary Astrophysics},
-         year = 2020,
-        month = nov,
-          eid = {arXiv:2011.12300},
-        pages = {arXiv:2011.12300},
+        title = "{Stratified and Vertically Shearing Streaming Instabilities in Protoplanetary Disks}",
+      journal = {\apj},
+     keywords = {Protoplanetary disks, Hydrodynamics, Astrophysical fluid dynamics, Planet formation, Planetesimals, 1300, 1963, 101, 1241, 1259, Astrophysics - Earth and Planetary Astrophysics},
+         year = 2021,
+        month = feb,
+       volume = {907},
+       number = {2},
+          eid = {64},
+        pages = {64},
+          doi = {10.3847/1538-4357/abcd9b},
 archivePrefix = {arXiv},
        eprint = {2011.12300},
  primaryClass = {astro-ph.EP},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv201112300L},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2021ApJ...907...64L},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
