### 2021-06-22 #3079 @jsoishi
@dpsanders I've done so. Thanks

On Sun, Jun 20, 2021 at 7:45 PM David P. Sanders ***@***.***>
wrote:

> @jsoishi <https://github.com/jsoishi>: Could you please check the issues
> with the DOIs signalled above. Thanks!
>
