### 2021-06-22 b519f9a
Added API changes notes
* for upgrading from version 1 to version 2
