### 2021-06-22 d875acd
Added note for API changes to README.md
