### 2021-06-22 #3079 @kyleniemeyer
Hi @jsoishi, I'm doing some final checks before publishing. Thanks for fixing those DOI issues. I still see one minor issue with the Lin reference: the journal name is missing. Can you please add that? Otherwise, this looks good to go.
