+++ JOSS-paper/paper.bib
@@ -235,7 +235,7 @@ archivePrefix = {arXiv},
 @ARTICLE{2021ApJ.907.64L,
        author = {{Lin}, Min-Kai},
         title = "{Stratified and Vertically Shearing Streaming Instabilities in Protoplanetary Disks}",
-      journal = {\apj},
+      journal = {Astrophysical Journal},
      keywords = {Protoplanetary disks, Hydrodynamics, Astrophysical fluid dynamics, Planet formation, Planetesimals, 1300, 1963, 101, 1241, 1259, Astrophysics - Earth and Planetary Astrophysics},
          year = 2021,
         month = feb,
