### 2021-06-23 #3079 @jsoishi
Hi @kyleniemeyer, thanks for catching that. I've pushed to the joss_paper branch with a fix for that just now.
